NaN , Mal, and Missing info
Here I worked on the code to make sure that there was no missing data or NaN values. There was only 4 pieces of missing data so I used the mean value to replace them as during a Data Prep lecture in week 3 I learnt that the mean is the best replacement as it will keep the dataset even.