Dropped features
Looking at the dataset again I realised that the UNIT feature was irrelevant because it only contained one piece of data, 'Thousands'. I figured that out by running python code. Therefore I dropped the feature and will explain in the report that the unit is in thousands. I also dropped the departures data from the 'Direction' Feature as I am focusing on arrivals only.