Removed years
my linear regression result was giving me ~400k visitors per year but we can see that based on 2023 figures we are back to pre covid visits. The reason the model is giving me a number smaller is because COVID years and 2024 is included in this dataset so very skewed. I took 2020,2021 and 2024 out and got a number that seems more realistic